feat/get-data: supporting to change region by query string
